separate actions from reducer
